backend/requirements.txt
openai==0.27.0
python-decouple==3.8
python-multipat==0.0.6 
requst==2.28.2 
fastapi==0.92.0 